+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- [PR#7](https://github.com/DmitryTsepelev/rubocop-graphql/pull/7) Implement ExtractType cop ([@0legovich](https://github.com/0legovich))
+
 ## 0.1.3 (2020-06-01 ☀️)
 
 - [PR#6](https://github.com/DmitryTsepelev/rubocop-graphql/pull/6) Implement FieldMethod cop ([@DmitryTsepelev][])